refactor(server): IDP module
